requirements.txt
%%writefile requirements.txt
streamlit
google-generativeai